NCS-506 PR adjustments
