+++ src/controllers/tasks/statement.of.capital.controller.ts
@@ -1,12 +1,13 @@
 import { NextFunction, Request, Response } from "express";
-import { RADIO_BUTTON_VALUE, STATEMENT_OF_CAPITAL_ERROR } from "../../utils/constants";
+import { RADIO_BUTTON_VALUE, sessionCookieConstants, STATEMENT_OF_CAPITAL_ERROR } from "../../utils/constants";
 import { STATEMENT_OF_CAPITAL_PATH, TASK_LIST_PATH, urlParams } from "../../types/page.urls";
 import { Templates } from "../../types/template.paths";
 import { urlUtils } from "../../utils/url";
 import { getStatementOfCapitalData } from "../../services/statement.of.capital.service";
 import { Session } from "@companieshouse/node-session-handler";
 import {
-  ConfirmationStatementSubmission, SectionStatus,
+  ConfirmationStatementSubmission,
+  SectionStatus,
   StatementOfCapital
 } from "private-api-sdk-node/dist/services/confirmation-statement";
 import { formatTitleCase } from "../../utils/format";
@@ -61,7 +62,7 @@ export const post = async (req: Request, res: Response, next: NextFunction) => {
 };
 
 const sendUpdate = async (transactionId: string, submissionId: string, req: Request) => {
-  const statementOfCapital: StatementOfCapital = req.sessionCookie["statementOfCapital"];
+  const statementOfCapital: StatementOfCapital = req.sessionCookie[sessionCookieConstants.STATEMENT_OF_CAPITAL_KEY];
   const session = req.session as Session;
   const csSubmission = buildCsSubmission(submissionId, transactionId, statementOfCapital, SectionStatus.CONFIRMED);
   await updateConfirmationStatement(session, transactionId, submissionId, csSubmission);
