+++ test/mocks/confirmation.statement.submission.mock.ts
@@ -1,20 +1,27 @@
-import { ConfirmationStatementSubmission, SectionStatus } from "private-api-sdk-node/dist/services/confirmation-statement";
+import {
+  ConfirmationStatementSubmission,
+  SectionStatus,
+  StatementOfCapital
+} from "private-api-sdk-node/dist/services/confirmation-statement";
+
+export const mockStatementOfCapital: StatementOfCapital = {
+  classOfShares: "Ordinary",
+  currency: "GBP",
+  numberAllotted: "100",
+  aggregateNominalValue: "0.01",
+  prescribedParticulars: "THE QUICK BROWN FOX",
+  totalNumberOfShares: "100",
+  totalAggregateNominalValue: "1",
+  totalAmountUnpaidForCurrency: "2"
+};
+
 
 export const mockConfirmationStatementSubmission: ConfirmationStatementSubmission = {
   id: "dgshjgdsj",
   data: {
     statementOfCapitalData: {
       sectionStatus: SectionStatus.CONFIRMED,
-      statementOfCapital: {
-        aggregateNominalValue: "12",
-        classOfShares: "class",
-        currency: "GBP",
-        numberAllotted: "123",
-        prescribedParticulars: "sssds",
-        totalAggregateNominalValue: "321.57",
-        totalAmountUnpaidForCurrency: "111",
-        totalNumberOfShares: "555"
-      }
+      statementOfCapital: mockStatementOfCapital
     }
   },
   links: {
